backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==3.2.2
cffi==2.0.0
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
gevent==25.9.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
websocket==0.2.1
zope.event==6.1
zope.interface==8.2
openai
